requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
pymysql==1.1.0
python-dotenv==0.19.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
email-validator==2.0.0
